+++ Changes.md
@@ -96,6 +96,7 @@
 - fix(outbound): replace empty Message-ID header #3288
 - fix(outbound): don't send SNI servername when connecting to an IP
 - fix(outbound): chown queue dir after creation #3291
+- fix(server): async endpoint.bind() and await in server.js #3366
 
 ### [3.0.3] - 2024-02-07